chore:extract BaseRepository to prepare for replacing the demo's API in the future.
chore:extract BaseRepository to prepare for replacing the demo's API in the future.

chore:extract BaseRepository to prepare for replacing the demo's API in the future.
