+++ sample/src/main/java/com/airwallex/paymentacceptance/repo/BaseRepository.kt
@@ -1,0 +1,17 @@
+package com.airwallex.paymentacceptance.repo
+
+import com.airwallex.android.core.model.PaymentIntent
+
+interface BaseRepository {
+
+    suspend fun login()
+
+    suspend fun getPaymentIntentFromServer(
+        force3DS: Boolean = false,
+        customerId: String? = null
+    ): PaymentIntent
+
+    suspend fun getCustomerIdFromServer(): String
+
+    suspend fun getClientSecretFromServer(customerId: String): String
+}
