+++ sample/src/main/java/com/airwallex/paymentacceptance/viewmodel/base/BaseViewModel.kt
@@ -20,6 +20,8 @@ import com.airwallex.paymentacceptance.Settings
 import com.airwallex.paymentacceptance.api.Api
 import com.airwallex.paymentacceptance.api.ApiFactory
 import com.airwallex.paymentacceptance.products
+import com.airwallex.paymentacceptance.repo.BaseRepository
+import com.airwallex.paymentacceptance.repo.LocalMockRepository
 import com.airwallex.paymentacceptance.shipping
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.Dispatchers
@@ -32,15 +34,6 @@ import java.util.UUID
 import java.util.concurrent.atomic.AtomicInteger
 
 abstract class BaseViewModel : ViewModel() {
-
-    private val api: Api
-        get() {
-            if (TextUtils.isEmpty(AirwallexPlugins.environment.baseUrl())) {
-                throw IllegalArgumentException("Base url should not be null or empty")
-            }
-            return ApiFactory(AirwallexPlugins.environment.baseUrl()).buildRetrofit()
-                .create(Api::class.java)
-        }
 
     private val _createPaymentIntentError = MutableLiveData<String>()
     val createPaymentIntentError: LiveData<String> = _createPaymentIntentError
@@ -57,14 +50,17 @@ abstract class BaseViewModel : ViewModel() {
 
     abstract fun init(activity: Activity)
 
+    private val repository: BaseRepository by lazy {
+        LocalMockRepository()
+    }
+
     /**
      * this method demonstrates how to log in using an apiKey and clientId.
      * this process should be completed on your own server;
      * do not copy this method.
      */
     private suspend fun login() {
-        val response = api.authentication(Settings.apiKey, Settings.clientId)
-        Settings.token = JSONObject(response.string())["token"].toString()
+        repository.login()
     }
 
     /**
@@ -76,35 +72,7 @@ abstract class BaseViewModel : ViewModel() {
         force3DS: Boolean = false,
         customerId: String? = null
     ): PaymentIntent {
-        if (customerId == null) {
-            login()
-        }
-        return withContext(Dispatchers.IO) {
-            val body = mutableMapOf(
-                "request_id" to UUID.randomUUID().toString(),
-                "amount" to Settings.price.toDouble(),
-                "currency" to Settings.currency,
-                "merchant_order_id" to UUID.randomUUID().toString(),
-                "order" to PurchaseOrder.Builder()
-                    .setProducts(products)
-                    .setShipping(shipping)
-                    .setType("physical_goods")
-                    .build()
-                    .toParamMap(),
-                "descriptor" to "Airwallex - T-sh  irt",
-                "metadata" to mapOf("id" to 1),
-                "email" to "yimadangxian@airwallex.com",
-                "return_url" to Settings.returnUrl
-            )
-            if (force3DS) {
-                body["payment_method_options"] =
-                    mapOf("card" to mapOf("three_ds_action" to "FORCE_3DS"))
-            }
-            Settings.cachedCustomerId?.let { body.put("customer_id", it) }
-            customerId?.let { body.put("customer_id", it) }
-            val paymentIntentResponse = api.createPaymentIntent(body)
-            PaymentIntentParser().parse(JSONObject(paymentIntentResponse.string()))
-        }
+        return repository.getPaymentIntentFromServer(force3DS, customerId)
     }
 
     /**
@@ -113,34 +81,7 @@ abstract class BaseViewModel : ViewModel() {
      * do not copy this method;instead, obtain the customerId from your own server.
      */
     suspend fun getCustomerIdFromServer(): String {
-        return withContext(Dispatchers.IO) {
-            login()
-            if (Settings.cachedCustomerId.isNullOrEmpty()) {
-                val customerResponse = api.createCustomer(
-                    mutableMapOf(
-                        "request_id" to UUID.randomUUID().toString(),
-                        "merchant_customer_id" to UUID.randomUUID().toString(),
-                        "first_name" to "John",
-                        "last_name" to "Doe",
-                        "email" to "john.doe@airwallex.com",
-                        "phone_number" to "13800000000",
-                        "additional_info" to mapOf(
-                            "registered_via_social_media" to false,
-                            "registration_date" to "2019-09-18",
-                            "first_successful_order_date" to "2019-09-18"
-                        ),
-                        "metadata" to mapOf(
-                            "id" to 1
-                        )
-                    )
-                )
-                val customerId = JSONObject(customerResponse.string())["id"].toString()
-                Settings.cachedCustomerId = customerId
-                customerId
-            } else {
-                Settings.cachedCustomerId!!
-            }
-        }
+        return repository.getCustomerIdFromServer()
     }
 
     /**
@@ -149,10 +90,7 @@ abstract class BaseViewModel : ViewModel() {
      * do not copy this method;instead, obtain the clientSecret from your own server.
      */
     suspend fun getClientSecretFromServer(customerId: String): String {
-        return withContext(Dispatchers.IO) {
-            val clientSecretResponse = api.createClientSecret(customerId)
-            ClientSecretParser().parse(JSONObject(clientSecretResponse.string())).value
-        }
+        return repository.getClientSecretFromServer(customerId)
     }
 
     suspend fun <T> loadPagedItems(
